minimum functional version of Here! read more for details
Users can now enter their valid UCD email and be able to check in to an
event.
Site admins can view the list of people who were at the event.
Form validation is done on both the client and the server side.

Form must:
- have one event selected
- have a valid UCD email. anything that has "ucdavis.edu" at the end of
it will be accepted.

Server checks:
- if email is valid by using the ucd.py function that searches for the
email in the UCD public directory. if email is not valid, it returns
the user to the form with a prompt that says the email is invalid.

Still need:
- a way to view all the events that a user has attended.
- count number of people that attended the event
- sign in profiles for users?